`mc-usb-device` & `mc-serial-port-device` catalogs decoupled
