+++ packages/pure/scripts/parse-request-configs.ts
@@ -1,3 +1,8 @@
+/**
+ * Copyright (c) Mudita sp. z o.o. All rights reserved.
+ * For licensing, see https://github.com/mudita/mudita-center/blob/master/LICENSE.md
+ */
+
 import { RequestConfig } from "../src"
 
 const parseRequestConfigs = (requestConfigsString = ""): RequestConfig[] => {
