+++ packages/pure/scripts/single-request.ts
@@ -1,4 +1,10 @@
-import { MuditaDevice, RequestConfig } from "../src"
+/**
+ * Copyright (c) Mudita sp. z o.o. All rights reserved.
+ * For licensing, see https://github.com/mudita/mudita-center/blob/master/LICENSE.md
+ */
+
+import { MuditaDevice } from "../src"
+import { RequestConfig } from "../src"
 
 const singleRequest = async (
   device: MuditaDevice,
