+++ packages/pure/src/device/base-device.ts
@@ -3,45 +3,42 @@
  * For licensing, see https://github.com/mudita/mudita-center/blob/master/LICENSE.md
  */
 
-import {
-  DeviceEventName,
-  McSerialPortDevice,
-  RequestConfig,
-  Response,
-  McUsbFile,
-  McUsbDevice,
-} from "./device.types"
+import { McSerialPortDeviceClass } from "../mc-serial-port-device/mc-serial-port-device.class"
+import { McUsbDeviceClass } from "../mc-usb-device/mc-usb-device.class"
+import { McUsbFile } from "../mc-usb-device/mc-usb-file.interface"
+import { RequestConfig, Response } from "../mc-serial-port-device/types"
+import { MuditaDevice, DeviceEventName } from "./mudita-device"
 
 export const timeoutMs = 30000
 
-class BaseDevice implements McSerialPortDevice, McUsbDevice {
+class BaseDevice implements MuditaDevice {
   constructor(
-    private baseMcSerialPortDevice: McSerialPortDevice,
-    private baseMcUsbDevice: McUsbDevice
+    private mcSerialPortDevice: McSerialPortDeviceClass,
+    private mcUsbDevice: McUsbDeviceClass
   ) {}
   get path() {
-    return this.baseMcSerialPortDevice.path
+    return this.mcSerialPortDevice.path
   }
   get deviceType() {
-    return this.baseMcSerialPortDevice.deviceType
+    return this.mcSerialPortDevice.deviceType
   }
   connect(): Promise<Response> {
-    return this.baseMcSerialPortDevice.connect()
+    return this.mcSerialPortDevice.connect()
   }
   disconnect(): Promise<Response> {
-    return this.baseMcSerialPortDevice.disconnect()
+    return this.mcSerialPortDevice.disconnect()
   }
   request(config: RequestConfig<any>): Promise<Response<any>> {
-    return this.baseMcSerialPortDevice.request(config)
+    return this.mcSerialPortDevice.request(config)
   }
   on(eventName: DeviceEventName, listener: () => void): void {
-    return this.baseMcSerialPortDevice.on(eventName, listener)
+    return this.mcSerialPortDevice.on(eventName, listener)
   }
   off(eventName: DeviceEventName, listener: () => void): void {
-    return this.baseMcSerialPortDevice.off(eventName, listener)
+    return this.mcSerialPortDevice.off(eventName, listener)
   }
   getFiles(): Promise<McUsbFile[]> {
-    return this.baseMcUsbDevice.getFiles()
+    return this.mcUsbDevice.getFiles()
   }
 }
