+++ packages/pure/src/device/index.ts
@@ -5,9 +5,9 @@
 
 export * from "./base-device"
 export * from "./device"
-export * from "./device.types"
 export * from "./constants"
 export * from "./descriptors"
 export * from "./services"
 export * from "./validators"
 export * from "./strategies"
+export * from "./mudita-device"
