+++ packages/pure/src/index.ts
@@ -5,10 +5,12 @@
 
 export * from "./device"
 export * from "./logger"
-export * from "./endpoints"
+export * from "./mc-serial-port-device/types"
 export * from "./formatter"
 export * from "./device-manager"
 export * from "./timeout"
+export * from "./mc-serial-port-device"
+export * from "./mc-usb-device"
 
 import PureNode from "./device-manager"
 export default PureNode
