+++ packages/pure/src/mc-serial-port-device/helpers/index.ts
@@ -1,0 +1,6 @@
+/**
+ * Copyright (c) Mudita sp. z o.o. All rights reserved.
+ * For licensing, see https://github.com/mudita/mudita-center/blob/master/LICENSE.md
+ */
+
+export * from "./is-api-request-payload"
