+++ packages/pure/src/mc-serial-port-device/index.ts
@@ -3,11 +3,5 @@
  * For licensing, see https://github.com/mudita/mudita-center/blob/master/LICENSE.md
  */
 
-class UsbDeviceService {
-  getObjectHandles() {
-    const handles = ["1", "2"]
-    return handles
-  }
-}
-
-export default UsbDeviceService
+export * from "./helpers"
+export * from "./types"
