+++ packages/pure/src/mc-serial-port-device/types/contact.types.ts
