+++ packages/pure/src/mc-serial-port-device/types/device-info.types.ts
