+++ packages/pure/src/mc-serial-port-device/types/index.ts
@@ -7,7 +7,9 @@ export * from "./backup.types"
 export * from "./contact.types"
 export * from "./device-info.types"
 export * from "./device-update.types"
-export * from "./file-system"
+export * from "./file-system.types"
 export * from "./messages.types"
 export * from "./outbox.types"
 export * from "./restore.types"
+export * from "./response.types"
+export * from "./request.types"
