+++ packages/pure/src/mc-serial-port-device/types/request.types.ts
@@ -1,0 +1,45 @@
+/**
+ * Copyright (c) Mudita sp. z o.o. All rights reserved.
+ * For licensing, see https://github.com/mudita/mudita-center/blob/master/LICENSE.md
+ */
+
+export enum Endpoint {
+  Invalid = 0,
+  DeviceInfo = 1,
+  Update = 2,
+  FileSystem = 3,
+  Backup = 4,
+  Restore = 5,
+  Factory = 6,
+  Contacts = 7,
+  Messages = 8,
+  CallLog = 9,
+  Security = 13,
+  Outbox = 14,
+
+  // api version (mocked)
+  ApiVersion = 1000,
+}
+
+export enum Method {
+  Get = 1,
+  Post = 2,
+  Put = 3,
+  Delete = 4,
+}
+
+export interface ApiRequestPayload extends RequestPayload {
+  endpoint: Endpoint.ApiVersion
+  method: Method.Get
+}
+
+export interface RequestPayload<T = undefined> extends RequestConfig<T> {
+  uuid: number
+}
+
+export interface RequestConfig<Body = undefined> {
+  endpoint: Endpoint
+  method: Method
+  body?: Body
+  filePath?: string
+}
