+++ packages/pure/src/mc-serial-port-device/types/response.types.ts
@@ -1,0 +1,45 @@
+/**
+ * Copyright (c) Mudita sp. z o.o. All rights reserved.
+ * For licensing, see https://github.com/mudita/mudita-center/blob/master/LICENSE.md
+ */
+
+import { Endpoint } from "./request.types"
+
+export enum ResponseStatus {
+  Ok = 200,
+  Accepted = 202,
+  Redirect = 303,
+  NoContent = 204,
+  BadRequest = 400,
+  NotFound = 404,
+  PhoneLocked = 403,
+  NotAcceptable = 406,
+  Conflict = 409,
+  InternalServerError = 500,
+  UnprocessableEntity = 422,
+
+  // lib status
+  ConnectionError = 503,
+  ParserError = 504,
+  Timeout = 505,
+}
+
+export type ResponseErrorCode = number
+
+interface ResponseError {
+  code?: ResponseErrorCode
+  message: string
+}
+
+export interface Response<Body = undefined> {
+  status: ResponseStatus
+  body?: Body
+  endpoint?: Endpoint
+  uuid?: number
+  error?: ResponseError
+}
+
+export interface PaginationBody {
+  limit: number
+  offset: number
+}
