+++ packages/pure/src/mc-usb-device/index.ts
@@ -1,0 +1,10 @@
+/**
+ * Copyright (c) Mudita sp. z o.o. All rights reserved.
+ * For licensing, see https://github.com/mudita/mudita-center/blob/master/LICENSE.md
+ */
+
+export * from "./mc-usb-device.class"
+export * from "./mc-usb-device"
+export * from "./mc-usb-file.interface"
+export * from "./usb-device.service"
+export * from "./usb-device.service.factory"
