+++ packages/pure/src/mc-usb-device/mc-usb-device.class.ts
@@ -1,0 +1,10 @@
+/**
+ * Copyright (c) Mudita sp. z o.o. All rights reserved.
+ * For licensing, see https://github.com/mudita/mudita-center/blob/master/LICENSE.md
+ */
+
+import { McUsbFile } from "./mc-usb-file.interface"
+
+export interface McUsbDeviceClass {
+  getFiles(): Promise<McUsbFile[]>
+}
