+++ packages/pure/src/mc-usb-device/mc-usb-file.interface.ts
@@ -1,0 +1,17 @@
+/**
+ * Copyright (c) Mudita sp. z o.o. All rights reserved.
+ * For licensing, see https://github.com/mudita/mudita-center/blob/master/LICENSE.md
+ */
+
+export enum McUsbFileType {
+  wav = 0x3008,
+  mp3 = 0x3009,
+  flac = 0xb906,
+}
+
+export interface McUsbFile {
+  id: string
+  size: number
+  name: string
+  type: McUsbFileType
+}
